# API/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Goomar API
  version: 1.0.0
  description: API for creating date/time/type entries in Goomar application

security:
  - bearerAuth: [ ]
servers:
  - url: http://localhost:9999
paths:
  /V1/token:
    post:
      tags:
        - Tokens
      summary: Generate bearer token
      operationId: getToken
      requestBody:
        required: true
        content:
          application/json; charset=utf-8:
            schema:
              $ref: '#/components/schemas/GetTokenReq'
      responses:
        '200':
          description: Token response
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/TokenRsp'

  /V1/appointments/{year}/{month}/{day}:
    get:
      tags:
        - Appointments
      summary: Get all application users
      operationId: getAppointments
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: path
          required: true
          schema:
            type: integer
        - name: day
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of application users
          content:
            application/json; charset=utf-8:
              schema:
                type: array
                default: [ ]
                items:
                  $ref: "#/components/schemas/ReservationRest"
  /V1/freeslots/{year}/{month}/{day}:
    get:
      tags:
        - Reservations
      summary: Get all application users
      operationId: getFreeSlots
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: path
          required: true
          schema:
            type: integer
        - name: day
          in: path
          required: true
          schema:
            type: integer
        - name: long
          in: query
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: List of application users
          content:
            application/json; charset=utf-8:
              schema:
                type: array
                default: [ ]
                items:
                  $ref: "#/components/schemas/FreeSlotRest"
  /V1/confirmation:
    get:
      tags:
        - Reservations
      summary: Get all application users
      operationId: getConfirmation
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of application users
          content:
            text/html:
              schema:
                type: string

  /V1/appointment:
    patch:
      tags:
        - Appointments
      summary: Delete user by Id
      operationId: confirmAppointment
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
        - name: eventId
          in: query
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Single user deleted
    delete:
      tags:
        - Appointments
      summary: Delete user by Id
      operationId: deleteAppointment
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
        - name: eventId
          in: query
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Single user deleted
  /V1/reservation:
    post:
      tags:
        - Reservations
      summary: Post reservation from public page
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json; charset=utf-8:
            schema:
              type: object
              $ref: '#/components/schemas/ReservationRest'
      responses:
        '201':
          description: create/update reservation
          content:
            application/json:
              schema:
                type: integer
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # optional, for documentation purposes only
  schemas:
    GetTokenReq:
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 5
          maxLength: 30
        password:
          type: string
          minLength: 5
          maxLength: 80
    TokenRsp:
      properties:
        token:
          type: string
    ReservationRest:
      required:
        - id
        - username
        - email
        - phone
        - date_time
        - registration
        - longService
      properties:
        id:
          type: integer
        username:
          type: string
          minLength: 5
          maxLength: 128
        date_time:
          type: string
          format: date-time
        email:
          type: string
          minLength: 5
          maxLength: 80
        phone:
          type: string
          minLength: 5
          maxLength: 80
        registration:
          type: string
          minLength: 5
          maxLength: 80
        event_id:
          type: string
          minLength: 5
          maxLength: 80
        longService:
          type: boolean
        confirmed:
          type: boolean
    FreeSlotRest:
      required:
        - start
        - end
      properties:
        start:
          type: string
          format: time
        end:
          type: string
          format: time